requirements.txt
Flask==2.3.2
python-nmap==0.7.1